+++ virtualizarr/tests/test_backend.py
@@ -1,3 +1,4 @@
+import functools
 from collections.abc import Mapping
 from concurrent.futures import ThreadPoolExecutor
 from pathlib import Path
@@ -17,7 +18,6 @@ from virtualizarr.backend import (
     automatically_determine_filetype,
 )
 from virtualizarr.manifests import ManifestArray
-from virtualizarr.parallel import DaskDelayedExecutor, LithopsEagerFunctionExecutor
 from virtualizarr.readers import HDF5VirtualBackend
 from virtualizarr.readers.hdf import HDFVirtualBackend
 from virtualizarr.tests import (
@@ -508,41 +508,67 @@ class TestLoadVirtualDataset:
             assert vds.scalar.attrs == {"scalar": "true"}
 
 
+preprocess_func = functools.partial(
+    xr.Dataset.rename_vars,
+    air="nair",
+)
+
+
 @requires_hdf5plugin
 @requires_imagecodecs
 @parametrize_over_hdf_backends
-@pytest.mark.parametrize(
-    "parallel",
-    [
-        False,
-        ThreadPoolExecutor,
-        pytest.param(DaskDelayedExecutor, marks=requires_dask),
-        pytest.param(
-            LithopsEagerFunctionExecutor,
-            marks=[
-                requires_lithops,
-                pytest.mark.xfail(
-                    reason="Lithops bug - see https://github.com/lithops-cloud/lithops/issues/1428"
-                ),
-            ],
-        ),
-    ],
-)
 class TestOpenVirtualMFDataset:
-    def test_parallel_open(self, netcdf4_files_factory, hdf_backend, parallel):
+    @pytest.mark.parametrize("invalid_parallel_kwarg", ["ray", Dataset])
+    def test_invalid_parallel_kwarg(
+        self, netcdf4_files_factory, invalid_parallel_kwarg, hdf_backend
+    ):
         filepath1, filepath2 = netcdf4_files_factory()
 
-        # test combine nested without in-memory indexes
+        with pytest.raises(ValueError, match="Unrecognized argument"):
+            open_virtual_mfdataset(
+                [filepath1, filepath2],
+                combine="nested",
+                concat_dim="time",
+                backend=hdf_backend,
+                parallel=invalid_parallel_kwarg,
+            )
+
+    @pytest.mark.parametrize(
+        "parallel",
+        [
+            False,
+            ThreadPoolExecutor,
+            pytest.param("dask", marks=requires_dask),
+            pytest.param("lithops", marks=requires_lithops),
+        ],
+    )
+    @pytest.mark.parametrize(
+        "preprocess",
+        [
+            None,
+            preprocess_func,
+        ],
+    )
+    def test_parallel_open(
+        self, netcdf4_files_factory, hdf_backend, parallel, preprocess
+    ):
+        filepath1, filepath2 = netcdf4_files_factory()
+        vds1 = open_virtual_dataset(filepath1, backend=hdf_backend)
+        vds2 = open_virtual_dataset(filepath2, backend=hdf_backend)
+
+        expected_vds = xr.concat([vds1, vds2], dim="time")
+        if preprocess:
+            expected_vds = preprocess_func(expected_vds)
+
+        # test combine nested, which doesn't use in-memory indexes
         combined_vds = open_virtual_mfdataset(
             [filepath1, filepath2],
             combine="nested",
             concat_dim="time",
             backend=hdf_backend,
             parallel=parallel,
+            preprocess=preprocess,
         )
-        vds1 = open_virtual_dataset(filepath1, backend=hdf_backend)
-        vds2 = open_virtual_dataset(filepath2, backend=hdf_backend)
-        expected_vds = xr.concat([vds1, vds2], dim="time")
         xrt.assert_identical(combined_vds, expected_vds)
 
         # test combine by coords using in-memory indexes
@@ -551,12 +577,7 @@ class TestOpenVirtualMFDataset:
             combine="by_coords",
             backend=hdf_backend,
             parallel=parallel,
-        )
-        vds1 = open_virtual_dataset(filepath1, backend=hdf_backend)
-        vds2 = open_virtual_dataset(filepath2, backend=hdf_backend)
-        expected_vds = xr.concat(
-            [vds1, vds2],
-            dim="time",
+            preprocess=preprocess,
         )
         xrt.assert_identical(combined_vds, expected_vds)
 
@@ -567,5 +588,6 @@ class TestOpenVirtualMFDataset:
             combine="by_coords",
             backend=hdf_backend,
             parallel=parallel,
+            preprocess=preprocess,
         )
-        xrt.assert_identical(combined_vds, expected_vds
+        xrt.assert_identical(combined_vds, expected_vds)